+++ madshifta/Source/UI/FilmStripKnob.cpp
@@ -50,7 +50,7 @@ void FilmStripKnob::mouseDrag(const juce::MouseEvent& event)
     float newValue = dragStartValue + deltaY * dragSensitivity;
     newValue = juce::jlimit(0.0f, 1.0f, newValue);
 
-    if (std::fabsf(newValue - value) > 1e-10f)
+    if (std::fabs(newValue - value) > 1e-10f)
     {
         value = newValue;
         repaint();
@@ -66,7 +66,7 @@ void FilmStripKnob::mouseUp(const juce::MouseEvent& /*event*/)
 void FilmStripKnob::setValue(float newValue)
 {
     newValue = juce::jlimit(0.0f, 1.0f, newValue);
-    if (std::fabsf(value - newValue) > 1e-10f)
+    if (std::fabs(value - newValue) > 1e-10f)
     {
         value = newValue;
         repaint();
