+++ supaphaser/Source/UI/FilmStripSlider.cpp
@@ -27,7 +27,7 @@ void FilmStripSlider::mouseDown(const juce::MouseEvent& event)
     // Set value directly based on click position
     float newValue = 1.0f - static_cast<float>(event.getPosition().y - minY) / static_cast<float>(maxY - minY);
     newValue = juce::jlimit(0.0f, 1.0f, newValue);
-    if (std::fabsf(newValue - value) > 1e-10f)
+    if (std::fabs(newValue - value) > 1e-10f)
     {
         value = newValue;
         repaint();
@@ -44,7 +44,7 @@ void FilmStripSlider::mouseDrag(const juce::MouseEvent& event)
 void FilmStripSlider::setValue(float newValue)
 {
     newValue = juce::jlimit(0.0f, 1.0f, newValue);
-    if (std::fabsf(value - newValue) > 1e-10f)
+    if (std::fabs(value - newValue) > 1e-10f)
     {
         value = newValue;
         repaint();
